Change version of parameterized models to 4.3
This commit modifies all the java docs and concerned files in previous
commits.
