+++ subprojects/tooling-api/src/crossVersionTest/groovy/org/gradle/integtests/tooling/r43/InvalidModel2Action.java
@@ -14,11 +14,18 @@
  * limitations under the License.
  */
 
-package org.gradle.integtests.tooling.r42;
+package org.gradle.integtests.tooling.r43;
 
-public interface CustomParameter {
+import org.gradle.api.Action;
+import org.gradle.tooling.BuildAction;
+import org.gradle.tooling.BuildController;
 
-    void setValue(String str);
-
-    String getValue();
+public class InvalidModel2Action implements BuildAction<CustomModel2> {
+    @Override
+    public CustomModel2 execute(BuildController controller) {
+        return controller.getModel(CustomModel2.class, CustomParameter.class, new Action<CustomParameter>() {
+            @Override
+            public void execute(CustomParameter customParameter) {}
+        });
+    }
 }
