+++ subprojects/tooling-api/src/main/java/org/gradle/tooling/internal/consumer/versioning/VersionDetails.java
@@ -36,8 +36,8 @@ public abstract class VersionDetails implements Serializable {
     }
 
     public static VersionDetails from(GradleVersion version) {
-        if (version.getBaseVersion().compareTo(GradleVersion.version("4.2")) >= 0) {
-            return new R42VersionDetails(version.getVersion());
+        if (version.getBaseVersion().compareTo(GradleVersion.version("4.3")) >= 0) {
+            return new R43VersionDetails(version.getVersion());
         }
         if (version.getBaseVersion().compareTo(GradleVersion.version("3.5")) >= 0) {
             return new R35VersionDetails(version.getVersion());
@@ -178,8 +178,8 @@ public abstract class VersionDetails implements Serializable {
         }
     }
 
-    private static class R42VersionDetails extends R35VersionDetails {
-        public R42VersionDetails(String version) {
+    private static class R43VersionDetails extends R35VersionDetails {
+        public R43VersionDetails(String version) {
             super(version);
         }
