+++ subprojects/tooling-api/src/main/java/org/gradle/tooling/internal/protocol/InternalCancellableConnection.java
@@ -24,13 +24,13 @@ import org.gradle.tooling.internal.protocol.exceptions.InternalUnsupportedBuildA
  *
  * <p>DO NOT CHANGE THIS INTERFACE - it is part of the cross-version protocol.
  *
- * <p>Consumer compatibility: This interface is used by all consumer versions from 2.1-rc-1 to 4.1. It is also used by later consumers when the provider does not
+ * <p>Consumer compatibility: This interface is used by all consumer versions from 2.1-rc-1 to 4.2. It is also used by later consumers when the provider does not
  * implement newer interfaces.
  * </p>
  * <p>Provider compatibility: This interface is implemented by all provider versions from 2.1-rc-1.</p>
  *
  * @since 2.1-rc-1
- * @deprecated 4.2. Use {@link InternalCancellableConnectionVersion2} instead.
+ * @deprecated 4.3. Use {@link InternalCancellableConnectionVersion2} instead.
  * @see org.gradle.tooling.internal.protocol.ConnectionVersion4
  */
 @Deprecated
@@ -38,7 +38,7 @@ public interface InternalCancellableConnection extends InternalProtocolInterface
     /**
      * Performs some action against a build and returns the requested model.
      *
-     * <p>Consumer compatibility: This method is used by all consumer versions from 2.1-rc-1 to 4.1. It is also used by later consumers when the provider does not
+     * <p>Consumer compatibility: This method is used by all consumer versions from 2.1-rc-1 to 4.2. It is also used by later consumers when the provider does not
      * implement newer interfaces.
      * </p>
      * <p>Provider compatibility: This method is implemented by all provider versions from 2.1-rc-1.</p>
@@ -51,7 +51,7 @@ public interface InternalCancellableConnection extends InternalProtocolInterface
      * @throws InternalBuildCancelledException When the operation was cancelled before it could complete.
      * @throws IllegalStateException When this connection has been stopped.
      * @since 2.1-rc-1
-     * @deprecated 4.2. Use {@link InternalCancellableConnectionVersion2#getModel(ModelIdentifier, InternalCancellationToken, BuildParameters)} instead.
+     * @deprecated 4.3. Use {@link InternalCancellableConnectionVersion2#getModel(ModelIdentifier, InternalCancellationToken, BuildParameters)} instead.
      */
     BuildResult<?> getModel(ModelIdentifier modelIdentifier, InternalCancellationToken cancellationToken,
                             BuildParameters operationParameters) throws
@@ -64,7 +64,7 @@ public interface InternalCancellableConnection extends InternalProtocolInterface
     /**
      * Performs some action against a build and returns the result.
      *
-     * <p>Consumer compatibility: This method is used by all consumer versions from 2.1-rc-1 to 4.1. It is also used by later consumers when the provider does not
+     * <p>Consumer compatibility: This method is used by all consumer versions from 2.1-rc-1 to 4.2. It is also used by later consumers when the provider does not
      * implement newer interfaces.
      * </p>
      * <p>Provider compatibility: This method is implemented by all provider versions from 2.1-rc-1.</p>
@@ -75,7 +75,7 @@ public interface InternalCancellableConnection extends InternalProtocolInterface
      * @throws InternalBuildCancelledException When the operation was cancelled before it could complete.
      * @throws IllegalStateException When this connection has been stopped.
      * @since 2.1-rc-1
-     * @deprecated 4.2. Use {@link InternalCancellableConnectionVersion2#run(InternalBuildActionVersion2, InternalCancellationToken, BuildParameters)} instead.
+     * @deprecated 4.3. Use {@link InternalCancellableConnectionVersion2#run(InternalBuildActionVersion2, InternalCancellationToken, BuildParameters)} instead.
      */
     <T> BuildResult<T> run(InternalBuildAction<T> action,
                            InternalCancellationToken cancellationToken,
